+++ src/components/RenderOptions.js
@@ -1,6 +1,7 @@
-import React, { Component } from 'react';
+import React, { Component } from 'react'
 import { TextField } from 'office-ui-fabric-react/lib/TextField'
-import { Slider } from 'office-ui-fabric-react/lib/Slider';
+import { Slider } from 'office-ui-fabric-react/lib/Slider'
+import { Label } from 'office-ui-fabric-react/lib/Label'
 
 export class RenderOptions extends Component {
     constructor() {
@@ -18,7 +19,7 @@ export class RenderOptions extends Component {
                 <div style={{ padding: '10px' }}>
                     <div className='flexColumns'>
                         <div style={{ flex: '1' }}>
-                            <TextField label='width'
+                            <TextField label='Width'
                                 value={this.props.items.width}
                                 onGetErrorMessage={this._getErrorMessage}
                                 onChange={(s, e) => this.props.onChange({
@@ -27,16 +28,25 @@ export class RenderOptions extends Component {
                                 })} />
                         </div>
                     </div>
-                    <Slider
-                        label="Scale"
-                        max={100}
-                        value={this.props.items.viewScale}
-                        showValue={true}
-                        onChange={(e) => this.props.onChange({
-                            width: this.props.items.width,
-                            viewScale: e
-                        })}
-                    />
+                    <div className='flexColumns'
+                        style={{ flexGrow: 'stretch', marginTop: '10px' }}>
+                        <div style={{width: 'calc(100% - 40px)'}}>
+                            <Slider label="Scale"
+                                max={100}
+                                value={this.props.items.viewScale}
+                                showValue={false}
+                                onChange={(e) => this.props.onChange({
+                                    width: this.props.items.width,
+                                    viewScale: e
+                                })}
+                            />
+                        </div>
+                        <div style={{marginTop: '18px', marginLeft: '5px'}}>
+                            <Label>{this.props.items.viewScale === 0
+                                ? 'Auto'
+                                : this.props.items.viewScale + '%'}</Label>
+                        </div>
+                    </div>
                 </div>
             }
         </div >